chore: refactor quotes
