+++ client-app/core/api/graphql/index.ts
@@ -8,7 +8,6 @@ export * from "./orders";
 export * from "./organization";
 export * from "./page";
 export * from "./payment";
-export * from "./quotes";
 export * from "./slugInfo";
 export * from "./store";
 export * from "./utils";
