+++ client-app/shared/account/composables/index.ts
@@ -7,5 +7,3 @@ export * from "./useUserCheckoutDefaults";
 export * from "./useUserOrder";
 export * from "./useUserOrders";
 export * from "./useUserOrdersFilter";
-export * from "./useUserQuote";
-export * from "./useUserQuotes";
